ggml-gobject: Add initial version of ggml-gobject
So far, we have bindings for some of the main primitives, eg,
tensors and the context. There is also some extra functionality to
enable the loading of models, for example the simple GPT2 model.

The main extras are GGMLModelDescNode/GGMLModelDescLeaf, which can
be used to describe the layout of tensors in a model. Then there is
GGMLTokenDictionary to provide an abstraction around a simple token-to-index
mapping and GGMLModel, which can be used to wrap some set of model weights.

To load a model from a .bin file, there's ggml_load_model_from_istream,
which basically replicates what the GPT2 demo does, but it uses GInputStream
to do it.