ggml-gobject: Add functions for loading the model and its parts async
And make them part of the public API.

This commit also changes a little bit of those APIs, namely that
instead of passing GGMLModelDescCreateFunc to ggml_model_load_from_istream,
we instead create the GGMLModelDesc as soon as possible (eg, once
we have the hyperparameters loaded) and then pass that directly to
ggml_model_load_from_istream.

Same also with n_tokens - we don't pass the entire GGMLHyperparameters
to ggml_load_token_dictionary_from_istream - we only need n_tokens
 so just pass that once it becomes available.

All this so that we can async load with callbacks from language
bindings - its not possible to put the entire loading process
in a separate thread since it needs to execute a callback
into js, and that needs to be on the main thread.