ggml-language-model: Only free the completion monitor state once everything is done
Its possible that calling g_task_return_pointer/error will require
another iteration of the main loop, so we have to keep the
GGMLLanguageModelCompleteMonitorState around until that's all finished -
we can do that with some refcounting.

(This can happen if, on the same mainloop iteration, we fetch both
the first result (eg, the prompt) back from the GAsyncQueue and then
an error, which will then cause the whole state (including the task's
user data) to be freed before it has a chance to actually run the task
callback handler on the next mainloop iteration).