llm-writer-app: Add completion states and handle more events
in STATE_TEXT_EDITOR, you can freely edit the text

In STATE_PREDICTING, we're waiting for predictions from the model,
so we display a spinner. Pressing backspace or key_left cancels
the prediction in progress.

In STATE_WAITING, we're waiting for user confirmation of the prediction.
They can either press key_right to accept the prediction, or key_left
to reject it.