ggml-cached-model: Add ggml_cached_model_istream_set_download_progress_callback
This uses GGMLProgressIstream to monitor the progress of the
GInputStream for the download, so that it can be displayed
in a friendly way to the user (instead of the read operation
just stalling forever).

It also means that ggml_cached_model_istream_new now returns
a GGMLCachedModelIstream as opposed to a GFileInputStream, so that
the user can invoke the set_download_progress_callback method.