llm-writer-app: Add selection of models and download feedback
When selecting a new model, we download it straight away - we
also start downloading the model on the first user interaction
which requests a completion. Because the download can take a while,
a progress bar is displayed in an overlay in the bottom right hand
corner.

The download doesn't get cancelled if the user cancels the prediction
operation, but it will get cancelled if they change the model type