ggml-model-config: Add GGMLModelConfig
This can be used to pre-configure how we would like the model to be
loaded. Right now it supports a configuration for quantization. A design
principle of GGMLModelConfig is that it can be used also when the
config is NULL - this is why the config_get functions return a gboolean
indicating whether there is any change from the default in that part of
the configuration and also handle the case where the config object is NULL