ggml-language-model: If using quantization, set up quantization in the ModelDescNode
We do this after loading the hyperparameters and starting to set up
the ModelDesc. The ModelDesc contains all the information about the types
of the tensors and since we can convert them now on the fly during loading,
this is the perfect place to edit once we know the desired quantization
configuration.