ggml-execution-memory: Add "alloc" and "recorder" modes
These are based on the newly added ggml_alloc context mode in
ggml, which use an allocator to specify graph memory layout as opposed
to the naive linear allocator. We can also use the allocator to compute
how much memory is actually required.

The general flow would be that you first run the forward pass with
worst-case inputs in "recorder" mode to compute a maximal memory usage
profile. The recorder mode sets tensor data addresses to a region
that doesn't exist in memory and also takes care to ensure that writes
to the tensor using the ggml-gobject API don't actually write anything
to memory.

Then afterwards, you can allocate a buffer of the required size, then
 use the allocator in "alloc" mode to create the same layout, this time
backed by a real buffer.

Using the alloc mode is fairly cheap, since it doesn't require
any system calls (all the memory is allocated upfront).